[FEATURE] Make it able to have the app classes in the __init__.py (now the default)! Implement backward compatible layer for all contrib apps.
