--- pyplanet/conf/app_template/app.py
@@ -1,5 +1,0 @@
-from pyplanet.apps.config import AppConfig
-
-
-class {{ camel_case_app_name }}Config(AppConfig):
-	name = '{{ app_name }}'
