+++ tests/apps/new_import_style/__init__.py
@@ -1,0 +1,5 @@
+from pyplanet.apps.config import AppConfig
+
+
+class NewImportStyle(AppConfig):
+	pass
